copy webassets to appdata
