+++ VisualChatGPTStudioShared/ToolWindows/Turbo/TerminalWindowTurboControl.xaml.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Drawing;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Text.Json;
@@ -703,295 +704,18 @@ namespace JeffPires.VisualChatGPTStudio.ToolWindows.Turbo
             var textColor = ((SolidColorBrush)Application.Current.Resources[VsBrushes.WindowTextKey]).Color;
             var backgroundColor = ((SolidColorBrush)Application.Current.Resources[VsBrushes.WindowKey]).Color;
             var gptBubbleColor = ((SolidColorBrush)Application.Current.Resources[VsBrushes.ScrollBarBackgroundKey]).Color;
-
             var codeBackgroundColor = Color.FromRgb(
                 (byte)Math.Max(0, backgroundColor.R - 20),
                 (byte)Math.Max(0, backgroundColor.G - 20),
                 (byte)Math.Max(0, backgroundColor.B - 20)
             );
 
-            var cssTextColor = ToCssColor(textColor);
-            var cssBackgroundColor = ToCssColor(backgroundColor);
-            var cssGptBackgroundColor = ToCssColor(gptBubbleColor);
-            var cssCodeBackgroundColor = ToCssColor(codeBackgroundColor);
-
-            var html2 = $$"""
-                          <!doctype html>
-                          <html>
-                          <head>
-                            <meta charset="utf-8"/>
-                            <meta http-equiv='X-UA-Compatible' content='IE=edge' />
-                            <style>
-                              body {
-                                  font-family: Segoe UI, Tahoma, Geneva, Verdana, sans-serif;
-                                  margin: 10px;
-                                  background-color: {{cssBackgroundColor}};
-                                  color: {{cssTextColor}};
-                              }
-                              p {
-                                  margin-top: 1px;
-                                  margin-bottom: 1px;
-                              }
-                              pre {
-                                  overflow-x: auto;
-                                  white-space: pre;
-                                  background: {{cssCodeBackgroundColor}};
-                                  color: {{cssTextColor}};
-                                  padding: 8px;
-                                  font-family: Consolas, 'Courier New', monospace;
-                                  font-size: 13px;
-                                  margin: 6px 0;
-                              }
-                              code {
-                                  font-family: Consolas, 'Courier New', monospace;
-                                  font-size: 13px;
-                                  color: {{cssTextColor}};
-                              }
-                              .copy-btn {
-                                  position: absolute;
-                                  right: 8px;
-                                  top: 6px;
-                                  width: 15px;
-                                  height: 15px;
-                                  background: none;
-                                  border: none;
-                                  padding: 0;
-                                  cursor: pointer;
-                                  outline: none;
-                              }
-                              .copy-btn img {
-                                  width: 15px;
-                                  height: 15px;
-                                  display: block;
-                              }
-                              summary {
-                                  cursor: pointer;
-                                  user-select: none;
-                                  margin - left: 8px;
-                                  color: gray;
-                              }
-                              details {
-                                  padding: 0 8px 8px 8px;
-                                  margin: 8px;
-                                  overflow: hidden;
-                                  transition: height .3s ease;
-                                  font-size: 14px;
-                                  color: gray;
-                              }
-                              details summary {
-                                  cursor: pointer;
-                                  transition: margin 150ms ease-out;
-                              }
-                              details[open] summary {
-                                  margin-bottom: 10px;
-                              }
-                              .msg { margin:.4rem 0; display:flex; }
-                              .user  { justify-content:flex-end; }
-                              .gpt   { justify-content:flex-start; }
-                              .bubble { max-width:85%; padding:.6rem .9rem; border-radius:1rem; box-shadow: 0px 0px 2px {{cssCodeBackgroundColor}}; }
-                              .user .bubble { background:{{cssGptBackgroundColor}}; color:{{cssTextColor}}; }
-                              .gpt  .bubble { background:{{cssGptBackgroundColor}}; color:{{cssTextColor}}; }
-
-                              pre{margin:0 !important;border-radius:0}
-                              .hljs{padding:1rem !important}
-
-                          .code-box{
-                            position:relative;
-                            margin:.6rem 0;
-                            border-radius:6px;
-                            overflow:hidden;
-                            background: {{cssCodeBackgroundColor}};
-                          }
-                          .code-box header{
-                            display:flex;
-                            align-items:center;
-                            justify-content:space-between;
-                            background:#2d2d2d;
-                            color:#ccc;
-                            padding:.3rem .5rem;
-                            font-size:12px;
-                            text-transform:uppercase;
-                          }
-                          .code-box header span{user-select:none}
-                          .code-box header button{
-                            margin-left:.4rem;
-                            padding:.2rem .5rem;
-                            border:none;
-                            border-radius:3px;
-                            background:#0e639c;
-                            color:#fff;
-                            cursor:pointer;
-                            font-size:11px;
-                          }
-                          .code-box header button:hover{background:#1177bb}
-                          .code-box pre{
-                            margin:0 !important;
-                            padding:1rem;
-                            overflow-x:auto;
-                          }
-                          .hljs{background:transparent !important}
-
-                          .mermaid-box{
-                            display:flex;
-                            justify-content:center;
-                            padding:1rem;
-                            background: {{cssCodeBackgroundColor}};
-                            border-radius:6px;
-                            margin:.6rem 0;
-                          }
-
-                            </style>
-                          </head>
-                          <body>
-                            <div id="chat"></div>
-
-                          <!-- Markdown -->
-                          <!-- TODO: move to local resource -->
-                          <script src="https://cdn.jsdelivr.net/npm/marked/marked.min.js"></script>
-                          <script src="https://cdn.jsdelivr.net/npm/marked-highlight/lib/index.umd.min.js"></script>
-
-                          <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.9.0/highlight.min.js"></script>
-                          <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.9.0/languages/csharp.min.js"></script>
-                          <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.9.0/languages/sql.min.js"></script>
-
-                          <!-- code theme -->
-                          <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.9.0/styles/github-dark.min.css">
-
-                          <!-- mermaid -->
-                          <script src="https://cdn.jsdelivr.net/npm/mermaid/dist/mermaid.min.js"></script>
-
-                          <script>
-                              const highlightExt = markedHighlight.markedHighlight({
-                                langPrefix: 'language-',
-                                langPrefix: 'hljs language-',
-                                highlight(code, lang, info) {
-                                    const language = hljs.getLanguage(lang) ? lang : 'plaintext';
-                                    return hljs.highlight(code, { language }).value;
-                                }
-                              });
-
-                              function rawCode(el){
-                                return el.getAttribute('data-raw') || el.textContent;
-                              }
-
-                              function buildCodeBlock(lang, highlightedHtml, raw){
-                                const id = 'cb-' + Math.random().toString(36).slice(2);
-                                const langUp = (lang || 'text').toUpperCase();
-                                return `
-                                  <div class="code-box">
-                                    <header>
-                                      <span>${langUp}</span>
-                                      <div>
-                                        <button onclick="sendCode('${id}','copy')">Copy</button>
-                                        <button onclick="sendCode('${id}','apply')">Apply</button>
-                                      </div>
-                                    </header>
-                                    <pre><code id="${id}" class="language-${lang}" data-raw="${raw.replace(/"/g,'&quot;')}">${highlightedHtml}</code></pre>
-                                  </div>`;
-                              }
-
-                              /* send messages to WebView2 */
-                              function sendCode(id, action){
-                                const raw = document.getElementById(id).getAttribute('data-raw');
-                                window.chrome?.webview?.postMessage({
-                                  action:action,
-                                  code:raw
-                                });
-                              }
-
-                              const renderer = new marked.Renderer();
-
-                              renderer.code = function ({ text, lang }) {
-                                  if (lang?.toLowerCase() === 'mermaid') {
-                                    const id = 'mmd-' + Math.random().toString(36).slice(2);
-                                    return `<div class="mermaid-box" id="${id}" data-mermaid="${text.replace(/"/g,'&quot;')}"></div>`;
-                                  }
-
-                                const highlighted = lang && hljs.getLanguage(lang)
-                                  ? hljs.highlight(text, { language: lang }).value
-                                  : hljs.highlightAuto(text).value;
-
-                                return buildCodeBlock(lang, highlighted, text);
-                              };
-
-                              marked.use({ renderer });
-
-                              mermaid.initialize({ startOnLoad: false, suppressErrorRendering: true, theme: 'dark' });
-
-                              /* --- Mermaid --- */
-                              function renderMermaid(){
-                                document
-                                  .querySelectorAll('.mermaid-box:not(.processed)')
-                                  .forEach(box => {
-                                    box.classList.add('processed');
-                                    const def = box.getAttribute('data-mermaid');
-                                    mermaid.render('svg-' + box.id, def)
-                                      .then(({ svg }) => {
-                                         box.innerHTML = svg;
-                                       })
-                                      .catch(err => { box.innerHTML = '<pre style="color:red">mermaid: ' + err + '</pre>'; })
-                                  });
-
-                              }
-
-                            /* --------- think + markdown --------- */
-                            function splitThink(text){
-                              const parts = [];
-                              const thinkRegex = /<think>([\s\S]*?)<\/think>/gi;
-                              let lastIdx = 0;
-                              let m;
-                              while((m = thinkRegex.exec(text)) !== null){
-                                if(m.index > lastIdx){
-                                  parts.push({type:'md', text:text.slice(lastIdx, m.index)});
-                                }
-                                parts.push({type:'think', text:m[1]});
-                                lastIdx = thinkRegex.lastIndex;
-                              }
-                              if(lastIdx < text.length) parts.push({type:'md', text:text.slice(lastIdx)});
-                              return parts;
-                            }
-
-                            function renderFrag(f){
-                              if(f.type === 'think'){
-                                return `
-                                  <details><summary>thinking…</summary>
-                                      ${f.text.replace(/</g,'&lt;').replace(/>/g,'&gt;')}
-                                  </details>`;
-                              }
-                              return marked.parse(f.text);
-                            }
-
-                            function addMsg(role, rawText){
-                              const wrap = document.createElement('div');
-                              wrap.className = 'msg ' + role;
-                              const bubble = document.createElement('div');
-                              bubble.className = 'bubble';
-                              bubble.innerHTML = splitThink(rawText).map(renderFrag).join('');
-                              wrap.appendChild(bubble);
-                              document.getElementById('chat').appendChild(wrap);
-                              scrollToBottom();
-                            }
-
-                            function updateLastGpt(rawText){
-                              const last = document.querySelector('#chat .gpt:last-child .bubble');
-                              if(last) last.innerHTML = splitThink(rawText).map(renderFrag).join('');
-                              else addMsg('gpt', rawText);
-                              scrollToBottom();
-                            }
-
-                            function clearChat(){
-                              document.getElementById('chat').innerHTML = '';
-                            }
-
-                            function scrollToBottom(){
-                              window.scrollTo(0, document.body.scrollHeight);
-                            }
-                          </script>
-                          </body>
-                          </html>
-                          """;
-
-            _webView?.CoreWebView2.NavigateToString(html2);
+            WebAsset.DeployTheme(
+                ToCssColor(textColor),
+                ToCssColor(backgroundColor),
+                ToCssColor(codeBackgroundColor),
+                ToCssColor(gptBubbleColor));
+            _webView?.CoreWebView2.Navigate(WebAsset.GetTurboPath());
         }
 
         /// <summary>
