+++ VisualChatGPTStudioShared/VisualChatGPTStudioShared.projitems
@@ -58,6 +58,7 @@
     </Compile>
     <Compile Include="$(MSBuildThisFileDirectory)ToolWindows\Turbo\ChatEntity.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ToolWindows\Turbo\ChatMessageSegment.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ToolWindows\Turbo\WebAsset.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Utils\BooleanToVisibilityConverter.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Utils\InverseBooleanConverter.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ToolWindows\Turbo\TerminalWindowTurbo.cs" />
