+++ WebAssets/TurboChat.html
@@ -1,0 +1,35 @@
+﻿<!doctype html>
+<html lang="en-US" translate="no">
+<head>
+    <meta charset="utf-8"/>
+    <meta http-equiv='X-UA-Compatible' content='IE=edge' />
+    <link rel="stylesheet" href="TurboChat.css"/>
+    <title></title>
+</head>
+<body>
+<div id="chat"></div>
+
+<!-- Markdown -->
+<!-- TODO: move to local resource -->
+<script src="https://cdn.jsdelivr.net/npm/marked/marked.min.js"></script>
+<script src="https://cdn.jsdelivr.net/npm/marked-highlight/lib/index.umd.min.js"></script>
+
+<script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.9.0/highlight.min.js"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.9.0/languages/csharp.min.js"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.9.0/languages/sql.min.js"></script>
+
+
+
+<!-- mermaid -->
+<script src="https://cdn.jsdelivr.net/npm/mermaid/dist/mermaid.min.js"></script>
+
+<!-- dynamic theme -->
+<link rel="stylesheet" href="theme.css"/>
+
+<!-- code theme -->
+<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.9.0/styles/github-dark.min.css">
+
+<script src="TurboChat.js"></script>
+
+</body>
+</html>
